Reorganize analysis scripts
