+++ HPOlibExperimentUtils/optimizer_settings.yaml
@@ -2,6 +2,7 @@
 # PURE RANDOMSEARCH################################################################################################### #
 randomsearch:
   optimizer: randomsearch
+  display_name: "random search"
 
 # DRAGONFLY DEFAULT ################################################################################################## #
 dragonfly_default:
@@ -19,51 +20,60 @@ smac_hb_eta_3:
   optimizer: smac_hb
   eta: 3
   max_stages: 5
+  display_name: "SMAC-HB"
 
 smac_sh_eta_3:
   optimizer: smac_sh
   eta: 3
   max_stages: 5
+  display_name: "SMAC-HB"
 
 smac_hb_eta_2:
   optimizer: smac_hb
   eta: 2
   max_stages: 5
+  display_name: "SMAC-SH (eta2)"
 
 smac_sh_eta_2:
   optimizer: smac_sh
   eta: 2
   max_stages: 5
+  display_name: "SMAC-SH (eta2)"
 
 # HPBANDSTER BOHB #################################################################################################### #
 hpbandster_bohb_eta_3:
   optimizer: hpbandster_bohb
   eta: 3
   max_stages: 5
+  display_name: "BOHB"
 
 hpbandster_bohb_eta_2:
   optimizer: hpbandster_bohb
   eta: 2
   max_stages: 5
+  display_name: "BOHB (eta2)"
 
 hpbandster_hb_eta_3:
   optimizer: hpbandster_hb
   eta: 3
   max_stages: 5
+  display_name: "HB"
 
 hpbandster_hb_eta_2:
   optimizer: hpbandster_hb
   eta: 2
   max_stages: 5
+  display_name: "HB (eta2)"
 
 hpbandster_sh_eta_3:
   optimizer: hpbandster_sh
   eta: 3
   max_stages: 5
+  display_name: "SH"
 
 hpbandster_sh_eta_2:
   optimizer: hpbandster_sh
   eta: 2
   max_stages: 5
-
+  display_name: "SH (eta2)"
 ...
